requirements.txt
torch==2.4.1+cpu
torchaudio==2.4.1+cpu
torchvision==0.19.1+cpu
--index-url https://download.pytorch.org/whl/cpu
faiss-cpu==1.8.0
sentence-transformers==3.1.1
streamlit==1.39.0
pymupdf==1.24.10
hnswlib==0.8.0
rank-bm25==0.2.2
openai==1.51.0
numpy==2.1.1
transformers==4.46.3
accelerate==1.1.1
scikit-learn==1.5.2
pandas==2.2.3


# Core
python-dotenv
numpy
pandas

# PDF Processing
PyMuPDF

# NLP / Embeddings
sentence-transformers
transformers
torch

# Vector Search
faiss-cpu

# Keyword Search
rank-bm25

# LLM API
openai

# UI
streamlit

# Utils
scikit-learn
tqdm

# Optional (better text processing)
nltk
spacy